Create README.md
This is a tracery exercise that mimics the poetic theorist, lyricist, and cultural critic, Bob Dylan. In typical Dylan prose, the subject is a singluar individual he is often talking about in relation to an anecdote, metaphor, or imagined time---past or future. He utilizes rich visuals and has a mastery of the English language, as so evidenced by his studying of latin, friendships with beat authors, and understanding of how to carve a new space in poetic traditions. If you don't immediately understand the line (lyric, stanza, commentary, etc.) that's ok because neither does he.